\name{internals}
\alias{internals}
\alias{maxRSS}
\alias{seg.lm.fit0.gamma}
\alias{print.jumpointsVar}
\title{Internal Functions}
\description{
Functions for internal use only, or not yet documented.
}
\usage{
seg.lm.fit0.gamma(y, Z, PSI, control, round=FALSE, print.level=0)
maxRSS(RSS, S, n)
\method{print}{jumpointsVar}(x, digits=max(3L, getOption("digits") - 3L), \ldots)
}
\keyword{internal}

