\name{internals}
\alias{internals}
\title{Internal Functions}
\description{
Functions for internal use only, or not yet documented.
}
\usage{
seg.lm.fit0.gamma(y, Z, PSI, control, round=FALSE, print.level=0)
max.rss(RSS, S, n)
%pen.MDL(id, n)
\method{print}{jumpointsVar}(x, digits=max(3L, getOption("digits") - 3L), \ldots)
}
\keyword{internal}

